// app/(portal)/trls/work/changeofcategory/types/changeofcategory-type.tsx
export type StatusType = 
| 'PENDING-CUSTOMER-ACTION'
| 'PENDING-SCREENING'
| 'PENDING-ASSESSMENT'
| 'MANAGER-REJECTED'
| 'MANAGER-APPROVED'
| 'RECOMMENDED-FOR-APPROVAL'
| 'RECOMMENDED-FOR-REJECTION'
| 'PENDING-ENDORSEMENT'
| 'ENDORSEMENT-COMPLETE';